Refactor : url 수정 및 조건문 추가

+++ src/main/java/com/fivemin/mzpc/service/admin/AdminFoodService.java
@@ -16,10 +16,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.persistence.EntityManager;
-import javax.transaction.Transactional;
 import java.io.*;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -51,32 +51,40 @@ public class AdminFoodService {
         this.amazonS3Client = amazonS3Client;
     }
 
+    @Transactional(readOnly = true)
     public List<FoodDto> getFoodList(String storeCode, boolean topping) {
         List<Food> foods = foodRepository.findByStoreCode(storeCode,topping);
         List<FoodDto> foodDtos = new ArrayList<>();
 
-        for ( Food food : foods) {
-            CategoryDto categoryDto = CategoryDto.builder()
-                    .idx(food.category.getIdx())
-                    .code(food.category.getCode())
-                    .name(food.category.getName())
-                    .build();
+        if (foods!=null) {
+            for (Food food : foods) {
+                CategoryDto categoryDto = CategoryDto.builder()
+                        .idx(food.category.getIdx())
+                        .code(food.category.getCode())
+                        .name(food.category.getName())
+                        .build();
 
-            FoodDto foodDto = FoodDto.builder()
-                    .idx(food.getIdx())
-                    .code(food.getCode())
-                    .name(food.getName())
-                    .price(food.getPrice())
-                    .picture("https://mzpc-s3-bucket.s3.ap-northeast-2.amazonaws.com/"+food.getPicture())
-                    .description(food.getDescription())
-                    .stock(food.getStock())
-                    .topping(food.isTopping())
-                    .categoryDto(categoryDto)
-                    .build();
-            foodDtos.add(foodDto);
+                FoodDto foodDto = FoodDto.builder()
+                        .idx(food.getIdx())
+                        .code(food.getCode())
+                        .name(food.getName())
+                        .price(food.getPrice())
+                        .picture("https://mzpc-s3-bucket.s3.ap-northeast-2.amazonaws.com/" + food.getPicture())
+                        .description(food.getDescription())
+                        .stock(food.getStock())
+                        .topping(food.isTopping())
+                        .categoryDto(categoryDto)
+                        .build();
+                foodDtos.add(foodDto);
+
+
+            }
+            return foodDtos;
+        } else {
+            thorwStoreError();
         }
 
-        return foodDtos;
+        return null;
     }
 
 
@@ -84,77 +92,85 @@ public class AdminFoodService {
         Category category = categoryRepository.findByName(categoryName);
         log.info("categoryIdx :{} ",category.getIdx());
         List<Food> foodList = foodRepository.findByCategoryIdx(category.getIdx());
-       List<FoodDto> foodDtoList = new ArrayList<>();
+        List<FoodDto> foodDtoList = new ArrayList<>();
 
-       log.info("foodList : {} ",foodList);
-
-        for(Food foods : foodList) {
-            FoodDto foodDto = FoodDto.builder()
-                    .idx(foods.getIdx())
-                    .code(foods.getCode())
-                    .name(foods.getName())
-                    .price(foods.getPrice())
-                    .picture(foods.getPicture())
-                    .description(foods.getDescription())
-                    .topping(foods.isTopping())
-                    .build();
-            foodDtoList.add(foodDto);
+        if (category!= null && foodList!=null) {
+            for (Food foods : foodList) {
+                FoodDto foodDto = FoodDto.builder()
+                        .idx(foods.getIdx())
+                        .code(foods.getCode())
+                        .name(foods.getName())
+                        .price(foods.getPrice())
+                        .picture(foods.getPicture())
+                        .description(foods.getDescription())
+                        .topping(foods.isTopping())
+                        .build();
+                foodDtoList.add(foodDto);
+            }
+            return foodDtoList;
+        } else {
+            thorwStoreError();
         }
 
-        log.info("foodDtoList : {}",foodDtoList);
-        return foodDtoList;
+        return null;
     }
 
     @Transactional
     public void addFood(FoodDto foodDto, MultipartFile foodPicture,String categoryCode) {
         Category category = categoryRepository.findByCode(categoryCode);
 
-        Food food = Food.builder()
-                .idx(foodDto.getIdx())
-                .code(makeCode())
-                .name(foodDto.getName())
-                .picture(foodPicture.getOriginalFilename())
-                .price(foodDto.getPrice())
-                .description(foodDto.getDescription())
-                .stock(foodDto.getStock())
-                .topping(foodDto.isTopping())
-                .category(category)
-                .build();
+        if (category!=null) {
+            Food food = Food.builder()
+                    .idx(foodDto.getIdx())
+                    .code(makeCode())
+                    .name(foodDto.getName())
+                    .picture(foodPicture.getOriginalFilename())
+                    .price(foodDto.getPrice())
+                    .description(foodDto.getDescription())
+                    .stock(foodDto.getStock())
+                    .topping(foodDto.isTopping())
+                    .category(category)
+                    .build();
 
-        foodRepository.save(food);
-        fileUpload(foodPicture, foodPicture.getOriginalFilename());
-    }
+            foodRepository.save(food);
+            fileUpload(foodPicture, foodPicture.getOriginalFilename());
+        }
+        else {
+            thorwCategoryError();
+        }
 
-    private String makeCode(){
-        LocalDateTime currentDateTime = LocalDateTime.now();
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("'F'HHMMyyyymmddss");
-        return currentDateTime.format(formatter);
     }
 
     public FoodDto getFoodList(String categoryCode, String foodCode) {
         Food food = foodRepository.findByCode(categoryCode,foodCode);
 
-        CategoryDto categoryDto = CategoryDto.builder()
-                .idx(food.getCategory().getIdx())
-                .code(food.getCategory().getCode())
-                .name(food.getCategory().getName())
-                .build();
+        if (food!=null) {
+            CategoryDto categoryDto = CategoryDto.builder()
+                    .idx(food.getCategory().getIdx())
+                    .code(food.getCategory().getCode())
+                    .name(food.getCategory().getName())
+                    .build();
 
-        FoodDto foodDto = FoodDto.builder()
-                .idx(food.getIdx())
-                .code(food.getCode())
-                .name(food.getName())
-                .picture(food.getPicture())
-                .price(food.getPrice())
-                .description(food.getDescription())
-                .stock(food.getStock())
-                .topping(food.isTopping())
-                .createdAt(food.getCreatedAt())
-                .updateAt(food.getUpdatedAt())
-                .categoryDto(categoryDto)
-                .build();
+            FoodDto foodDto = FoodDto.builder()
+                    .idx(food.getIdx())
+                    .code(food.getCode())
+                    .name(food.getName())
+                    .picture(food.getPicture())
+                    .price(food.getPrice())
+                    .description(food.getDescription())
+                    .stock(food.getStock())
+                    .topping(food.isTopping())
+                    .createdAt(food.getCreatedAt())
+                    .updateAt(food.getUpdatedAt())
+                    .categoryDto(categoryDto)
+                    .build();
 
-        return foodDto;
+            return foodDto;
+        } else {
+            thorwFoodError();
+        }
+        return null;
+
     }
 
     @Transactional
@@ -163,22 +179,26 @@ public class AdminFoodService {
         Food updateFood = foodRepository.findById(foodDto.getIdx()).orElse(null);
         String fileName = foodPicture==null? updateFood.getPicture() : foodPicture.getOriginalFilename();
 
-        if (foodPicture != null){
-            deleteFile(updateFood.getPicture());
-            fileUpload(foodPicture,fileName);
+        if (category!=null && updateFood != null) {
+            if (foodPicture != null) {
+                deleteFile(updateFood.getPicture());
+                fileUpload(foodPicture, fileName);
+            }
+
+            updateFood.setIdx(foodDto.getIdx());
+            updateFood.setCode(foodDto.getCode());
+            updateFood.setName(foodDto.getName());
+            updateFood.setPrice(foodDto.getPrice());
+            updateFood.setPicture(fileName);
+            updateFood.setDescription(foodDto.getDescription());
+            updateFood.setStock(foodDto.getStock());
+            updateFood.setTopping(foodDto.isTopping());
+            updateFood.setCategory(category);
+
+            entityManager.merge(updateFood);
+        } else {
+            thorwFoodError();
         }
-
-        updateFood.setIdx(foodDto.getIdx());
-        updateFood.setCode(foodDto.getCode());
-        updateFood.setName(foodDto.getName());
-        updateFood.setPrice(foodDto.getPrice());
-        updateFood.setPicture(fileName);
-        updateFood.setDescription(foodDto.getDescription());
-        updateFood.setStock(foodDto.getStock());
-        updateFood.setTopping(foodDto.isTopping());
-        updateFood.setCategory(category);
-
-        entityManager.merge(updateFood);
 
     }
 
@@ -186,9 +206,14 @@ public class AdminFoodService {
     public void deleteFood(Long foodIdx) {
 
         Food food = foodRepository.findByIdx(foodIdx);
-        foodRepository.deleteById(foodIdx);
-        log.info("food_picture : {}",food.getPicture());
-        deleteFile(food.getPicture());
+        if (food != null) {
+            foodRepository.deleteById(foodIdx);
+            log.info("food_picture : {}", food.getPicture());
+            deleteFile(food.getPicture());
+        } else {
+            thorwFoodError();
+        }
+
     }
 
     public String fileUpload(MultipartFile foodPicture, String fileName) {
@@ -208,17 +233,10 @@ public class AdminFoodService {
                     uploadFile.delete();
                 }
             }
+
         }
         return returnResult;
-    }
 
-    //MutipartFile을 File로 타입 변경 메서드
-    private File convertMultiPartFileToFile(MultipartFile multipartFile) throws IOException {
-        File file = new File(multipartFile.getOriginalFilename());
-        try (FileOutputStream fos = new FileOutputStream(file)) {
-            fos.write(multipartFile.getBytes());
-        }
-        return file;
     }
 
     public void deleteFile(String foodPicture) {
@@ -233,9 +251,35 @@ public class AdminFoodService {
         } catch (Exception e) {
             System.err.println("파일 삭제 중 오류가 발생했습니다: " + e.getMessage());
         }
+
     }
 
+    private void thorwStoreError() {
+        throw new IllegalArgumentException("Store not found");
+    }
 
+    private void thorwCategoryError() {
+        throw new IllegalArgumentException("Category not found");
+    }
 
+    private void thorwFoodError() {
+        throw new IllegalArgumentException("Food not found");
+    }
+
+    private String makeCode(){
+        LocalDateTime currentDateTime = LocalDateTime.now();
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("'F'HHMMyyyymmddss");
+        return currentDateTime.format(formatter);
+    }
+
+    //MutipartFile을 File로 타입 변경 메서드
+    private File convertMultiPartFileToFile(MultipartFile multipartFile) throws IOException {
+        File file = new File(multipartFile.getOriginalFilename());
+        try (FileOutputStream fos = new FileOutputStream(file)) {
+            fos.write(multipartFile.getBytes());
+        }
+        return file;
+
+    }
 
 }
